Lazer finished Bloom not working on lazer Changed some damage & health values from int to float so that lazer can damage over time

+++ Assets/_Scripts/Enemies/Enemy.cs
@@ -10,7 +10,7 @@ public class Enemy : MonoBehaviour, IHittable, IAgent
     public EnemyDataSO EnemyData { get; set; }
 
     [field: SerializeField]
-    public int Health { get; set; }
+    public float Health { get; set; }
 
     [field: SerializeField]
     public int Damage { get; private set; }
@@ -53,7 +53,7 @@ public class Enemy : MonoBehaviour, IHittable, IAgent
         //blood = FindComponentInChildWithTag
     }
 
-    public void GetHit(int damage, GameObject damageDealer)
+    public void GetHit(float damage, GameObject damageDealer)
     {
         Health -= damage;
         blood.Play();
